requirements.txt
numpy>=1.20,<2
pillow>=9.0
torch>=2.0
tqdm>=4.0
